Release v1.4.2: Playtime tracking and real-time status updates

+++ src/renderer/duplicate-stack-overlay.js
@@ -14,7 +14,7 @@ export function createDuplicateStackOverlayController({
 
             const card = createCard(game, {
                 draggable: false,
-                launchMode: 'single',
+                launchMode: 'double',
                 showDuplicateChip: false,
                 showPath: false
             });
